Refactoring source code to make easy testing
