+++ package.json
@@ -40,7 +40,7 @@
     "imagemin-svgo": "^10.0.1",
     "iso-639-1": "^2.1.15",
     "iso-639-3": "^3.0.1",
-    "jest": "^29.3.1",
+    "jest": "^29.5.0",
     "js-yaml": "^4.1.0",
     "md5": "^2.3.0",
     "ncp": "^2.0.0",
@@ -50,7 +50,7 @@
     "snyk": "^1.1066.0",
     "sweetalert2": "^11.6.15",
     "tiny-async-pool": "^1.3.0",
-    "ts-jest": "^29.0.3",
+    "ts-jest": "^29.0.5",
     "ts-node": "^10.9.1",
     "typescript": "^4.9.3",
     "winston": "^3.8.2",
@@ -68,10 +68,10 @@
   },
   "scripts": {
     "setup": "ts-node-esm -- steps/setup.ts",
-    "get": "ts-node-esm -- steps/get.ts",
-    "transform": "ts-node-esm -- steps/transform.ts",
+    "get": "ts-node-esm -- steps/get/index.ts",
+    "transform": "ts-node-esm -- steps/transform/index.ts",
     "export-prebuild": "npm run build && node_modules/browserify/bin/cmd.js res/js/index.js -t babelify -o res/js/dist.js",
-    "export": "npm run export-prebuild && ts-node-esm -- steps/export.ts",
+    "export": "npm run export-prebuild && ts-node-esm -- steps/export/index.ts",
     "start": "npm run setup && npm run get && npm run transform && npm run export",
     "test": "npm run export-prebuild && node --experimental-vm-modules node_modules/jest/bin/jest.js --verbose --runInBand --forceExit",
     "build": "tsc",
