+++ tsconfig.json
@@ -19,8 +19,5 @@
     },
     "files": [
         "res/js/index.ts",
-        "steps/get.ts",
-        "steps/transform.ts",
-        "steps/export.ts"
     ]
 }
